split integrations_server files to read and payload
