+++ ec/ecresource/deploymentresource/integrationsserver/v2/integrations_server_payload.go
@@ -22,9 +22,6 @@ import (
 
 	"github.com/elastic/cloud-sdk-go/pkg/models"
 	topologyv1 "github.com/elastic/terraform-provider-ec/ec/ecresource/deploymentresource/topology/v1"
-	"github.com/elastic/terraform-provider-ec/ec/ecresource/deploymentresource/utils"
-	"github.com/elastic/terraform-provider-ec/ec/internal/converters"
-	"github.com/elastic/terraform-provider-ec/ec/internal/util"
 	"github.com/hashicorp/terraform-plugin-framework/diag"
 	"github.com/hashicorp/terraform-plugin-framework/tfsdk"
 	"github.com/hashicorp/terraform-plugin-framework/types"
@@ -42,77 +39,6 @@ type IntegrationsServerTF struct {
 	SizeResource              types.String `tfsdk:"size_resource"`
 	ZoneCount                 types.Int64  `tfsdk:"zone_count"`
 	Config                    types.Object `tfsdk:"config"`
-}
-
-type IntegrationsServer struct {
-	ElasticsearchClusterRefId *string                   `tfsdk:"elasticsearch_cluster_ref_id"`
-	RefId                     *string                   `tfsdk:"ref_id"`
-	ResourceId                *string                   `tfsdk:"resource_id"`
-	Region                    *string                   `tfsdk:"region"`
-	HttpEndpoint              *string                   `tfsdk:"http_endpoint"`
-	HttpsEndpoint             *string                   `tfsdk:"https_endpoint"`
-	InstanceConfigurationId   *string                   `tfsdk:"instance_configuration_id"`
-	Size                      *string                   `tfsdk:"size"`
-	SizeResource              *string                   `tfsdk:"size_resource"`
-	ZoneCount                 int                       `tfsdk:"zone_count"`
-	Config                    *IntegrationsServerConfig `tfsdk:"config"`
-}
-
-func ReadIntegrationsServers(in []*models.IntegrationsServerResourceInfo) (*IntegrationsServer, error) {
-	for _, model := range in {
-		if util.IsCurrentIntegrationsServerPlanEmpty(model) || utils.IsIntegrationsServerResourceStopped(model) {
-			continue
-		}
-
-		srv, err := readIntegrationsServer(model)
-		if err != nil {
-			return nil, err
-		}
-
-		return srv, nil
-	}
-
-	return nil, nil
-}
-
-func readIntegrationsServer(in *models.IntegrationsServerResourceInfo) (*IntegrationsServer, error) {
-
-	var srv IntegrationsServer
-
-	srv.RefId = in.RefID
-
-	srv.ResourceId = in.Info.ID
-
-	srv.Region = in.Region
-
-	plan := in.Info.PlanInfo.Current.Plan
-
-	topologies, err := readIntegrationsServerTopologies(plan.ClusterTopology)
-
-	if err != nil {
-		return nil, err
-	}
-
-	if len(topologies) > 0 {
-		srv.InstanceConfigurationId = topologies[0].InstanceConfigurationId
-		srv.Size = topologies[0].Size
-		srv.SizeResource = topologies[0].SizeResource
-		srv.ZoneCount = topologies[0].ZoneCount
-	}
-
-	srv.ElasticsearchClusterRefId = in.ElasticsearchClusterRefID
-
-	srv.HttpEndpoint, srv.HttpsEndpoint = converters.ExtractEndpoints(in.Info.Metadata)
-
-	cfg, err := readIntegrationsServerConfigs(plan.IntegrationsServer)
-
-	if err != nil {
-		return nil, err
-	}
-
-	srv.Config = cfg
-
-	return &srv, nil
 }
 
 func (srv IntegrationsServerTF) Payload(ctx context.Context, payload models.IntegrationsServerPayload) (*models.IntegrationsServerPayload, diag.Diagnostics) {
@@ -164,7 +90,7 @@ func IntegrationsServerPayload(ctx context.Context, srvObj types.Object, templat
 		return nil, nil
 	}
 
-	templatePayload := integrationsServerResource(template)
+	templatePayload := payloadFromTemplate(template)
 
 	if templatePayload == nil {
 		diags.AddError("integrations_server payload error", "integrations_server specified but deployment template is not configured for it. Use a different template if you wish to add integrations_server")
@@ -180,9 +106,9 @@ func IntegrationsServerPayload(ctx context.Context, srvObj types.Object, templat
 	return payload, nil
 }
 
-// integrationsServerResource returns the IntegrationsServerPayload from a deployment
+// payloadFromTemplate returns the IntegrationsServerPayload from a deployment
 // template or an empty version of the payload.
-func integrationsServerResource(template *models.DeploymentTemplateInfoV2) *models.IntegrationsServerPayload {
+func payloadFromTemplate(template *models.DeploymentTemplateInfoV2) *models.IntegrationsServerPayload {
 	if template == nil || len(template.DeploymentTemplate.Resources.IntegrationsServer) == 0 {
 		return nil
 	}
